refactor(api): move archive usecase to bounded context
