+++ api/tests/prescription/campaign/unit/domain/usecases/archive-campaigns_test.js
@@ -1,5 +1,5 @@
-import { archiveCampaigns } from '../../../../../lib/domain/usecases/campaigns-administration/archive-campaigns.js';
-import { expect, sinon } from '../../../../test-helper.js';
+import { archiveCampaigns } from '../../../../../../src/prescription/campaign/domain/usecases/archive-campaigns.js';
+import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | archive-campaign', function () {
   let campaignAdministrationRepository;
